requirements.txt
streamlit==1.54.0
pandas==2.3.3
plotly==6.5.2
langchain==1.2.10
langchain-google-genai==4.2.0
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-core==1.2.13
python-dotenv==1.2.1
google-generativeai==0.8.6
